Move export-modal to its own component folder
