+++ src/components/export-modal/export-modal.scss
@@ -1,0 +1,5 @@
+.pipeline-icon-modal {
+  .kui-button {
+    margin: 0 10px;
+  }
+}
