+++ src/components/icon-toolbar/icon-toolbar.scss
@@ -98,9 +98,3 @@
     height: 2em;
   }
 }
-
-.pipeline-icon-modal {
-  .kui-button {
-    margin: 0 10px;
-  }
-}
